requirement.txt
rest_framework
pymysql
django-cors-headers
celery-with-redis
django-celery
python-crontab